feat: Add rounding to hidden and visible features
Feature generation was not entirely deterministic on different machines
for different users. The difference was probably due to some rounding
errors, because the delta between the generated features as very-very
small in all cases. By adding an explicit rounding to the hidden and the
visible feature matrix, the feature generation can be made fully
deterministic.

Also: log computed hash on stdout for immediate checking.